requirements.txt
# ArcGIS Knowledge Integration - Python Dependencies

# HTTP requests
requests>=2.31.0

# Data processing
pydantic>=2.5.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code quality
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0

# Docker
docker>=7.0.0

# Environment
python-dotenv>=1.0.0